Moved index.js to root
Changes:
* Updated index.js contains error handling code by default and needs to
replace the existing index.js in the root
	>  Why? It was in the wrong location when it was copied over
        from the greater Pipelines repo

Future Work...
*
